Add files via upload
Added a new script for estimating the significance of cross-correlation between two parameters in a subset of filopodia in a larger dataset ("top correlating subcluster", TCS). The new script generates a large number of randomisations for one of the two parameters (tip movement) and computes cross-correlations of all the randomised datasets with the original tip fluorescence dataset for each filopodium.  

To recapitulate the hierarchical cluster analysis done in the original parent script, this script re-clusters each randomised dataset in the same way. This way, we generate as many "top correlating subclusters" from randomised datasets as desired. (500 take several minutes on my machine.) 

In essence, comparing the original TCS to these randomised TCSs informs us of the likelihood of finding the observed cross correlation between the two parameters in a subcluster of the same size purely by chance if the two parameters were uncoupled from each other.