Add files via upload
Downstream of Bounder_CCF.R. The aim of this script is to find whether the positive cross-correlation in top-correlating subclusters observed in our time series datasets is significantly above what is expected to result from the top-correlating subcluster from hierarchical clustering of a non-correlated (randomised) dataset.

The script generates a large number of randomisations of the movement parameter for each structure and re-calculates the cross-correlation between fluorescence and randomised movement. The CCFs of randomised data are re-clustered, and those randomisations are retained where the size of the top-correlating subcluster corresponds to the size of the top-correlating subcluster in the original non-randomised dataset (within 15% tolerance), with other randomisations rejected. The script also plots the mean cross-correlation values for the top correlating subcluster of all randomisations, compared with that of the original.